Add an option for the damping function
This PR adds the `damping` parameter to the foundation model calculator, which defaults to "bj" for backwards compatibility (and because it makes the most sense). This is also done for added clarity, as it's not mentioned anywhere in the docstring whether a damping function is used. In fact, at a quick glance, the user might be misled to thinking no damping is used since "D3" is printed to screen even though it is actually D3(BJ).